+++ receiver/otlpreceiver/internal/errors/errors.go
@@ -44,26 +44,3 @@ func GetHTTPStatusCodeFromStatus(s *status.Status) int {
 		return http.StatusInternalServerError
 	}
 }
-
-func NewStatusFromMsgAndHTTPCode(errMsg string, statusCode int) *status.Status {
-	var c codes.Code
-	// Mapping based on https://github.com/grpc/grpc/blob/master/doc/http-grpc-status-mapping.md
-	// 429 mapping to ResourceExhausted and 400 mapping to StatusBadRequest are exceptions.
-	switch statusCode {
-	case http.StatusBadRequest:
-		c = codes.InvalidArgument
-	case http.StatusUnauthorized:
-		c = codes.Unauthenticated
-	case http.StatusForbidden:
-		c = codes.PermissionDenied
-	case http.StatusNotFound:
-		c = codes.Unimplemented
-	case http.StatusTooManyRequests:
-		c = codes.ResourceExhausted
-	case http.StatusBadGateway, http.StatusServiceUnavailable, http.StatusGatewayTimeout:
-		c = codes.Unavailable
-	default:
-		c = codes.Unknown
-	}
-	return status.New(c, errMsg)
-}
